Move tests to proper directory
